Added some memory manager code

+++ src/include/stdlib.h
@@ -14,8 +14,8 @@ typedef signed char int8;
 typedef unsigned short uint16;
 typedef signed short int16;
 
-typedef unsigned long uint32;
-typedef signed long int32;
+typedef unsigned int uint32;
+typedef signed int int32;
 
 typedef unsigned long long uint64;
 typedef signed long long int64;
